Moving SquareShapedViewModifier to the HUD folder.
